# services/api/openapi.yml
openapi: 3.0.3
info:
  title: Healthy Diet API
  version: 1.0.0
  description: API specification synchronized with current Axum routes.
servers:
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      summary: Login
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginPayload'
      responses:
        '200':
          description: Login success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid email or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/admin/login:
    post:
      tags: [Authentication]
      summary: Login as admin/operator
      operationId: adminLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginPayload'
      responses:
        '200':
          description: Admin login success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid email or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Authenticated user is not an admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenPayload'
      responses:
        '200':
          description: New token pair
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid/expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/chat_check:
    get:
      tags: [Chat]
      summary: Check downstream chat agent availability
      description: Probes the configured downstream agent `/ping` route and reports whether proxy chat is currently reachable.
      operationId: chatCheck
      responses:
        '200':
          description: Agent is reachable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProxyChatCheckResponse'
        '500':
          description: Missing configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProxyChatCheckResponse'
        '503':
          description: Agent unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProxyChatCheckResponse'
  /api/chat:
    post:
      tags: [Chat]
      summary: Proxy chat request to agent service (SSE)
      description: Rust-facing SSE endpoint for frontend clients. This route persists local room/history metadata, then forwards the request to the downstream agent route `/api/chat`.
      operationId: chat
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentChatRequest'
      responses:
        '200':
          description: Server-Sent Events stream
          content:
            text/event-stream:
              schema:
                type: string
                example: "data: {\"type\":\"answer\",\"content\":\"...\"}\n\n"
        '400':
          description: Invalid image format/payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server or downstream error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/approve:
    post:
      tags: [Admin]
      summary: Approve or reject downstream agent action
      description: Completes approval workflows initiated by the downstream agent chat flow. This route proxies the agent route `/api/approve`.
      operationId: approveAgent
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentApproveRequest'
      responses:
        '200':
          description: Pass-through JSON from downstream service
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server/downstream error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/announcements/current:
    get:
      tags: [Knowledge]
      summary: Get the latest published public announcement
      operationId: getCurrentAnnouncement
      responses:
        '200':
          description: Current published announcement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentAnnouncementResponse'
        '404':
          description: No published announcement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/news/sync:
    post:
      tags: [Knowledge]
      summary: Trigger downstream FDA news sync
      description: Rust-facing proxy for the downstream agent route `/api/news/sync`.
      operationId: newsSync
      responses:
        '200':
          description: News sync result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsSyncResponse'
        '500':
          description: Server configuration error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Downstream service unavailable or invalid response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/news:
    get:
      tags: [Knowledge]
      summary: Get paginated local news list
      description: Rust-facing proxy for the downstream agent route `/api/news`.
      operationId: listNews
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            format: int64
            minimum: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            format: int64
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: News list response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsListResponse'
        '500':
          description: Server configuration error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Downstream service unavailable or invalid response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/knowledge-graph/status:
    get:
      tags: [Knowledge]
      summary: Get global knowledge graph readiness and counts
      description: Public Rust-facing endpoint for frontend clients. This route proxies the downstream agent route `/api/graph/status` and returns global graph readiness plus aggregate counts.
      operationId: knowledgeGraphStatus
      responses:
        '200':
          description: Knowledge graph global status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeGraphStatusResponse'
        '502':
          description: Downstream service unavailable or invalid response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/knowledge-graph/nodes:
    get:
      tags: [Knowledge]
      summary: List knowledge graph nodes from the global graph cache
      description: Public Rust-facing endpoint for frontend clients. This route proxies the downstream agent route `/api/graph/nodes`.
      operationId: knowledgeGraphNodes
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            format: int64
            minimum: 1
            maximum: 500
            default: 100
        - in: query
          name: node_type
          schema:
            type: string
        - in: query
          name: query
          schema:
            type: string
      responses:
        '200':
          description: Knowledge graph node list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeGraphNodesResponse'
        '502':
          description: Downstream service unavailable or invalid response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    AuthResponse:
      type: object
      required: [token, refreshToken, expiresIn, user]
      properties:
        token:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          format: int64
        user:
          $ref: '#/components/schemas/UserProfile'
...
    LoginPayload:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    RefreshTokenPayload:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
...
    ProxyChatCheckResponse:
      type: object
      required: [pingUrl, pingOk, proxyChatAvailable]
      properties:
        pingUrl:
          type: string
        pingOk:
          type: boolean
        proxyChatAvailable:
          type: boolean
        pingStatusCode:
          type: integer
          nullable: true
        pingResponse:
          type: string
          nullable: true
...
    CurrentAnnouncementResponse:
      type: object
      required: [id, title, content, publishedAt]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        content:
          type: string
        publishedAt:
          type: string
          format: date-time
...
    AgentApproveRequest:
      type: object
      required: [approval_id, action]
      properties:
        approval_id:
          type: string
        action:
          type: string
          enum: [approve, reject]
...